+++ R/fct_data_validation.R
@@ -1,0 +1,3 @@
+need_xlsx <- function(x) {
+  shiny::need(x == "xlsx", "Please, upload an xlsx file.")
+}