+++ NEWS.md
@@ -1,6 +1,9 @@
-# grouper (development version)
+# grouper 0.1.0
+
+* activate basic UI for the Shiny app.
+* defined functions to generate the tables with turns and rooms groups
 
 # grouper 0.0.0.9000
 
-* Addede basic development support
+* Added basic development support
 * Added a `NEWS.md` file to track changes to the package.